+++ test/radar/logger_test.rb
@@ -1,0 +1,11 @@
+require 'test_helper'
+
+class LoggerTest < Test::Unit::TestCase
+  context "logger class" do
+    setup do
+      @klass = Radar::Logger
+    end
+
+    # TODO. This is a tough one to test.
+  end
+end